# docs/swagger-v1.json
{
    "openapi":  "3.0.4",
    "info":  {
                 "title":  "SmartTaskManagement.API",
                 "version":  "1.0"
             },
    "paths":  {
...
                  "/api/projects":  {
                                        "get":  {
                                                    "tags":  [
                                                                 "Projects"
                                                             ],
                                                    "summary":  "Lists projects visible to the current user, with optional search, sorting, and pagination.",
                                                    "parameters":  [
                                                                       {
                                                                           "name":  "Search",
                                                                           "in":  "query",
                                                                           "schema":  {
                                                                                          "type":  "string"
                                                                                      }
                                                                       },
                                                                       {
                                                                           "name":  "SortField",
                                                                           "in":  "query",
                                                                           "schema":  {
                                                                                          "$ref":  "#/components/schemas/ProjectSortFieldDto"
                                                                                      }
                                                                       },
                                                                       {
                                                                           "name":  "SortDirection",
                                                                           "in":  "query",
                                                                           "schema":  {
                                                                                          "$ref":  "#/components/schemas/SortDirection"
                                                                                      }
                                                                       },
                                                                       {
                                                                           "name":  "PageNumber",
                                                                           "in":  "query",
                                                                           "schema":  {
                                                                                          "type":  "integer",
                                                                                          "format":  "int32"
                                                                                      }
                                                                       },
                                                                       {
                                                                           "name":  "PageSize",
                                                                           "in":  "query",
                                                                           "schema":  {
                                                                                          "type":  "integer",
                                                                                          "format":  "int32"
                                                                                      }
                                                                       }
                                                                   ],
                                                    "responses":  {
                                                                      "200":  {
                                                                                  "description":  "OK"
                                                                              }
                                                                  }
                                                },
                                        "post":  {
                                                     "tags":  [
                                                                  "Projects"
                                                              ],
                                                     "summary":  "Creates a new project. The owner is set to the authenticated caller.",
                                                     "requestBody":  {
                                                                         "description":  "Project creation input.",
                                                                         "content":  {
                                                                                         "application/json":  {
                                                                                                                  "schema":  {
                                                                                                                                 "$ref":  "#/components/schemas/CreateProjectRequestDto"
                                                                                                                             }
                                                                                                              },
                                                                                         "text/json":  {
                                                                                                           "schema":  {
                                                                                                                          "$ref":  "#/components/schemas/CreateProjectRequestDto"
                                                                                                                      }
                                                                                                       },
                                                                                         "application/*+json":  {
                                                                                                                    "schema":  {
                                                                                                                                   "$ref":  "#/components/schemas/CreateProjectRequestDto"
                                                                                                                               }
                                                                                                                }
                                                                                     }
                                                                     },
                                                     "responses":  {
                                                                       "200":  {
                                                                                   "description":  "OK"
                                                                               }
                                                                   }
                                                 }
                                    },
                  "/api/projects/{id}":  {
                                             "get":  {
                                                         "tags":  [
                                                                      "Projects"
                                                                  ],
                                                         "summary":  "Retrieves a single project by id.",
                                                         "parameters":  [
                                                                            {
                                                                                "name":  "id",
                                                                                "in":  "path",
                                                                                "description":  "Project identifier.",
                                                                                "required":  true,
                                                                                "schema":  {
                                                                                               "type":  "string",
                                                                                               "format":  "uuid"
                                                                                           }
                                                                            }
                                                                        ],
                                                         "responses":  {
                                                                           "200":  {
                                                                                       "description":  "OK"
                                                                                   }
                                                                       }
                                                     },
                                             "put":  {
                                                         "tags":  [
                                                                      "Projects"
                                                                  ],
                                                         "summary":  "Updates an existing project. Only the name and description are mutable.",
                                                         "parameters":  [
                                                                            {
                                                                                "name":  "id",
                                                                                "in":  "path",
                                                                                "description":  "Project identifier.",
                                                                                "required":  true,
                                                                                "schema":  {
                                                                                               "type":  "string",
                                                                                               "format":  "uuid"
                                                                                           }
                                                                            }
                                                                        ],
                                                         "requestBody":  {
                                                                             "description":  "Update input.",
                                                                             "content":  {
                                                                                             "application/json":  {
                                                                                                                      "schema":  {
                                                                                                                                     "$ref":  "#/components/schemas/UpdateProjectRequestDto"
                                                                                                                                 }
                                                                                                                  },
                                                                                             "text/json":  {
                                                                                                               "schema":  {
                                                                                                                              "$ref":  "#/components/schemas/UpdateProjectRequestDto"
                                                                                                                          }
                                                                                                           },
                                                                                             "application/*+json":  {
                                                                                                                        "schema":  {
                                                                                                                                       "$ref":  "#/components/schemas/UpdateProjectRequestDto"
                                                                                                                                   }
                                                                                                                    }
                                                                                         }
                                                                         },
                                                         "responses":  {
                                                                           "200":  {
                                                                                       "description":  "OK"
                                                                                   }
                                                                       }
                                                     },
                                             "delete":  {
                                                            "tags":  [
                                                                         "Projects"
                                                                     ],
                                                            "summary":  "Soft-deletes a project. The project and its tasks are hidden from normal queries.",
                                                            "parameters":  [
                                                                               {
                                                                                   "name":  "id",
                                                                                   "in":  "path",
                                                                                   "description":  "Project identifier.",
                                                                                   "required":  true,
                                                                                   "schema":  {
                                                                                                  "type":  "string",
                                                                                                  "format":  "uuid"
                                                                                              }
                                                                               }
                                                                           ],
                                                            "responses":  {
                                                                              "200":  {
                                                                                          "description":  "OK"
                                                                                      }
                                                                          }
                                                        }
                                         },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "CreateProjectRequestDto":  {
                                                                       "type":  "object",
                                                                       "properties":  {
                                                                                          "name":  {
                                                                                                       "type":  "string",
                                                                                                       "nullable":  true
                                                                                                   },
                                                                                          "description":  {
                                                                                                              "type":  "string",
                                                                                                              "nullable":  true
                                                                                                          }
                                                                                      },
                                                                       "additionalProperties":  false
                                                                   },
...
                                       "ProjectSortFieldDto":  {
                                                                   "enum":  [
                                                                                0,
                                                                                1,
                                                                                2
                                                                            ],
                                                                   "type":  "integer",
                                                                   "format":  "int32"
                                                               },
...
                                       "SortDirection":  {
                                                             "enum":  [
                                                                          0,
                                                                          1
                                                                      ],
                                                             "type":  "integer",
                                                             "format":  "int32"
                                                         },
...
                                       "UpdateProjectRequestDto":  {
                                                                       "type":  "object",
                                                                       "properties":  {
                                                                                          "name":  {
                                                                                                       "type":  "string",
                                                                                                       "nullable":  true
                                                                                                   },
                                                                                          "description":  {
                                                                                                              "type":  "string",
                                                                                                              "nullable":  true
                                                                                                          }
                                                                                      },
                                                                       "additionalProperties":  false
                                                                   },
...
                                   },
                       "securitySchemes":  {
                                               "Bearer":  {
                                                              "type":  "http",
                                                              "description":  "Enter the JWT access token (without the \u0027Bearer \u0027 prefix).",
                                                              "scheme":  "bearer",
                                                              "bearerFormat":  "JWT"
                                                          }
                                           }
                   },
    "security":  [
                     {
                         "Bearer":  [

                                    ]
                     }
                 ],
...
}